Move Heartbeat to taker
We only use this there so no need to have this in `daemon-shared`.
