+++ shared-bin/src/to_sse_event.rs
@@ -98,54 +98,3 @@ impl ToSseEvent for Option<Quote> {
         Event::json(self).event("quote")
     }
 }
-
-#[derive(Debug, Clone, Copy, Serialize)]
-pub struct Heartbeat {
-    timestamp: Timestamp,
-    interval: u64,
-}
-
-impl Heartbeat {
-    pub fn new(interval: u64) -> Self {
-        Self {
-            timestamp: Timestamp::now(),
-            interval,
-        }
-    }
-}
-
-impl ToSseEvent for Heartbeat {
-    fn to_sse_event(&self) -> Event {
-        Event::json(self).event("heartbeat")
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_test::Token;
-
-    #[test]
-    fn heartbeat_serialization() {
-        let heartbeat = Heartbeat {
-            timestamp: Timestamp::new(0),
-            interval: 1,
-        };
-
-        serde_test::assert_ser_tokens(
-            &heartbeat,
-            &[
-                Token::Struct {
-                    name: "Heartbeat",
-                    len: 2,
-                },
-                Token::Str("timestamp"),
-                Token::NewtypeStruct { name: "Timestamp" },
-                Token::I64(0),
-                Token::Str("interval"),
-                Token::U64(1),
-                Token::StructEnd,
-            ],
-        );
-    }
-}
